Refactor AJAX call into a service
This commit makes a class httpService with a class method for requesting the top word from the database and appending it to the DOM.
